Refactor quick edit panel to a separate package
